+++ src/pages/LoginPage.tsx
@@ -55,6 +55,7 @@ export const LoginPage: React.FC = () => {
     <div
       style={{
         minHeight: "100vh",
+        width: "100%",
         display: "flex",
         alignItems: "center",
         justifyContent: "center",